server_vps/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Werkzeug==3.0.1
requests==2.31.0
python-dotenv==1.0.0
openai==0.28.0
nvdlib==0.7.6